requirements.txt
absl-py==1.0.0
# actionlib==1.13.2
ale-py==0.7.5
angles==1.9.13
astor==0.8.1
asttokens==2.0.8
astunparse==1.6.3
atari-py==0.2.9
attrs==21.2.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
bondpy==1.8.6
cachetools==4.2.4
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
#certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
cffi==1.15.1
charset-normalizer==2.0.9
click==8.1.3
cloudpickle==1.6.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.2
Cython==0.29.32
decorator==5.1.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
fasteners==0.18
gast==0.2.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
glfw==2.5.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gym==0.21.0
gym-notices==0.0.8
h5py==2.10.0
idna==3.3
image-geometry==1.16.2
imageio==2.9.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
interactive-markers==1.12.0
joblib==1.1.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
laser_geometry==1.6.7
magent==0.1.14
Markdown==3.3.6
message-filters==1.15.14
mindspore @ https://ms-release.obs.cn-north-4.myhuaweicloud.com/1.8.1/MindSpore/cpu/x86_64/mindspore-1.8.1-cp37-cp37m-linux_x86_64.whl
mpi4py==3.1.3
mujoco-py==2.1.2.14
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.3
Pillow==8.2.0
pluggy==1.0.0
progressbar2==3.55.0
protobuf==3.19.1
psutil==5.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
pyglet==1.5.15
pyparsing==3.0.6
pytest==6.2.5
python-qt-binding==0.4.4
python-utils==2.5.6
PyWavelets==1.2.0
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.26.0
requests-oauthlib==1.3.0
resource_retriever==1.12.7
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.8
rviz==1.14.19
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
tensorboard==1.15.0
tensorboard-plugin-wit==1.8.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorlayer==2.2.3
termcolor==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
topic-tools==1.15.14
torch==1.12.1+cpu
torch-scatter==2.0.9
torchaudio==0.12.1+cpu
torchvision==0.13.1+cpu
tqdm==4.62.3
# typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
xacro==1.14.13
# zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
